move user avatar to a separate file
prop-8
