+++ app/react/components/PropsList/UserAvatar.jsx
@@ -1,0 +1,13 @@
+import React, { PropTypes } from 'react';
+
+const UserAvatar = ({ avatarPath }) => (
+  <img
+    className="praised-person-avatar"
+    src={avatarPath}
+    alt="avatar"
+  />
+);
+
+UserAvatar.propTypes = {
+  avatarPath: PropTypes.string,
+};
